// back_end/src/coupon/enum/coupon-message.enum.ts
export enum CouponMessage {
  MESSAGE_001 = 'CouponGlobals Code has existed!',
  MESSAGE_002 = 'CouponGlobals is not exist',
  MESSAGE_003 = 'CouponGlobals is deleted!',
  MESSAGE_004 = 'The coupon has expired!',
  MESSAGE_005 = 'The user has added this coupon.',
  MESSAGE_006 = 'This coupon has been distributed to this store.',
  MESSAGE_007 = 'The store does not exist.',
  MESSAGE_008 = 'The user has no coupons.',
  MESSAGE_009 = 'The coupon does not exist or has been used.',
}